refactor with http adapter
